WIP, move animateFOV to cesiumLIB, always restore zoom in screenspacecontroller when leaving wheel fov mode
